+++ changelog.md
@@ -3,3 +3,9 @@
 ## Functions
 - ✅ Fixed `ytdlp.download()` leaving some files with no extension
 - ✅ Fixed `rbuttonPrem().movePlayhead()` still throwing under certain circumstances
+
+## KSA
+- ✏️ Added `closePanel`
+
+## PremiereRemote
+- ✏️ Added `matchSelectedClipsToLowestTrack()` as a `UXP` function